Updates!
***Frontpage.html is the new home page not website.html. I started again because I wasn't getting anywhere with it.***

I've almost completed the home page, I just need someone to link the other pages to it once they're finished.

I've also done the snack page but I've come across a problem where the navbar is to big and I don't know how to change it. If you do, please do have a try.
I've checked the responsiveness of the websites and they are not very responsive, so if you have finished your page, please have a go to make it as responsive as possible.
I've changed the theme and fonts so I'll update the other pages as well.